requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==3.2.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
clarabel==0.11.1
click==8.3.1
cryptography==46.0.6
curl_cffi==0.13.0
cvxpy==1.8.2
distro==1.9.0
fastapi==0.135.2
filetype==1.2.0
frozendict==2.4.7
google-auth==2.49.1
google-genai==1.68.0
greenlet==3.3.2
h11==0.16.0
highspy==1.13.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.22
langchain-google-genai==4.2.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
Mako==1.3.10
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.4.3
orjson==3.11.7
ormsgpack==1.12.2
osqp==1.1.1
packaging==26.0
pandas==3.0.1
passlib==1.7.4
peewee==4.0.2
pgvector==0.4.2
platformdirs==4.9.4
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pyportfolioopt==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
requests-toolbelt==1.0.0
scikit-base==0.13.1
scikit-learn==1.8.0
scipy==1.17.1
scs==3.2.11
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
tenacity==9.1.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yfinance==1.2.0
zstandard==0.25.0
